Add files via upload
As part of the analysis for the online store specializes in selling video games globally, we will work with comprehensive dataset, encompassing user and expert reviews, genres, platforms (e.g. Xbox or PlayStation), and historical game sales data. The goal of this project is to uncover key patterns that contribute to a game's success. We need to identify potential high-performing titles and strategically plan effective advertising campaigns for the upcoming year.